Tkinter GUI Coffee Shop/requirements.txt
pillow==10.2.0
tkmacosx==1.0.5

